Updated PUT
On the client side PUT was modified to append the file name and size to the data sent. On the server side the program now parses this and uses the information to form a new file of that name, store the received data, and then close the file. The server will also now allow multiple PUT commands in a row without throwing an "address already in use" error.